Updated API. Updated copyrights. Added frame provider API.

+++ trunk/mp2dec2/lwmovie_layer2.hpp
@@ -1,5 +1,28 @@
+/*
+ * Copyright (c) 2014 Eric Lasota
+ * 
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ * 
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ * 
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
 #ifndef __LWMOVIE_LAYER2_HPP__
 #define __LWMOVIE_LAYER2_HPP__
+
+#include <ctype.h>
 
 typedef long long lwmSInt64;
 typedef int lwmSInt32;
@@ -7,7 +30,7 @@ typedef unsigned int lwmUInt32;
 typedef unsigned short lwmUInt16;
 typedef unsigned char lwmUInt8;
 typedef short lwmSInt16;
-typedef unsigned int lwmLargeUInt;
+typedef size_t lwmLargeUInt;
 
 typedef unsigned int lwmFastUInt8;
 typedef unsigned int lwmFastUInt16;
